chore: Fixing tests
